requirements.txt
Flask==3.1.1
Flask_SocketIO==5.5.1
python-dotenv==1.1.1